Adding an upscaling context id
This is to battle the strangeness of jessica/graphite.
There we will always get 100x100 display resolution, which is incorrect.
So now we have an upscaling context id, which is returned when we create a context, set in a render step which stores it in Tr2Renderer for easy retrieval.

https://ccpgames.atlassian.net/browse/PLAT-5646