Fixing a DeviceHung during ReleaseDeviceResources
Fixing an error where the fallback textures would be destroyed  while in use in a constant buffer.
Also fixed devicehung during enabling/disabling of the fallback texture debugging

The error was caused by us destroying the fallback textures during any device resource release (instead of only the video memory one) and recreated during the preparedeviceresource.
This would happen while the fallback textures were in use.
To fix this I removed the destruction of the fallback textures during releasedeviceresource (since the Tr2TextureAl will be released automatically when the memory is released) and added a check around the texture creation so we only create them if they are invalid.